Chapter 1 (Lexing) -> This is my Lexer for the monkey programming language
This is a brand new project for me, I have decided to learn about interpreters and how they operate. I didn't have the opportunity to do a module in college about this topic, so I decided that I wanted to learn how they work. I have bought: "Writing an interpreter in go by Thorsten Ball", and I'm working through his book to get an understand on how these systems work.

The purpose of a Lexer is to do the following:

This is where we take our source code and transform it to tokens (“Lexing”). Then the tokens are fed into a parser, which turns the tokens into an Abstract Syntax Tree (AST).