Today (20/10/25): I worked on: functions & function calls (the last part of the Evaluation chapter.) We refactored our "Eval" function in our evaluator.go file to take in an Environment (this is from environment.go). We wrote some more unit tests for our evaluator, and we also added our own "Function" type in our object package.
Yesterday (19/10/25) and Today (20/10/25): I worked on the following sections left in the Evaluation chapter:

bindings & the environment - I did this on the 19th of October!

functions & function calls - I did this on the 20th of October (today!)

this is the end of the evaluation chapter! Now I will go back to pratt parsing and try my best to understand it better now that I have completed the evaluation chapter!