Today (27/10/25), I started the final chapter of the book: "Extending the interpreter".
Below is a summary of what I have done today:

so far we have added in the following new types to the Monkey programming language:

strings & string concatenation
added in our own built-in-function: len, we will be adding in more.
array

I got to the top of page 223 today.

we did some small adding to some of our unit tests and we also wrote some new unit tests for our parser.

we did some refactoring in our parser.go file, and we also wrote some new functions.

we added some new types to our object system: String, Builtin, and Array.

we extended out lexer unit test, to now lex: strings and arrays.

in our lexer.go file we did some small changes, this ensure that we can lex: strings and arrays.

in our evaluator unit tests we did some small changes here and there, and we also added in some more unit tests.

in our evaluator.go file we did some refactoring and wrote some new functions.

in our evaluator package we have a new .go file this is called builtins.go, this is used for our built-in-functions for our interpreter (at the moment we only have "len", but this will be getting added to)!

and finally, we added some new node types for our AST, this is then used in our parser to make objects. Below are the new node types that we have added to our AST:

StringLiteral
ArrayLiteral
IndexExpression