Today (31/10/25), I finished the book: Writing an Interpreter in Go!
Today I have officially completed the book! I am sad that we are now done the book but also happy because now I get to begin a new project!

Today we did the following:

We added more built-in functions for arrays in our interpreter.

we added support for hashes: for example, {"name": "Kyle", "age": 22}

the grand finale - adding support for printing in our interpreter!

And, now we are done the book!

We wrote 3 more unit tests for parsing our new hash type.

We added some more code to our parser.go file to be able to parser our new hash type.

We made a brand new test file, for our object system.

We added some new types of our own to our object system.

We extended our lexer unit test.

We added a new case branch for our "NextToken" method to be able to lex a colon.

We added to our evaluator unit tests & we also wrote 2 new unit tests.

We did some refactoring in our Eval(...) function, evalIndexExpression, and we also wrote 2 new functions.

We added more built-in functions for arrays in our interpreter:

first
last
rest
push
puts

And finally we added to our ast.go file to now have a HashLiteral type, this is used for our new hash type, for exmpale: {"name": "Kyle", "age": 22}